+++ x/rewards/keeper/params.go
@@ -27,23 +27,6 @@ func (k Keeper) MinimumPriceOfGas(ctx sdk.Context) sdk.DecCoin {
 
 // GetParams return all module parameters.
 func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	store := ctx.KVStore(k.state.key)
-	bz := store.Get(types.ParamsKey)
-	if bz == nil {
-		return params
-	}
-
-	k.cdc.MustUnmarshal(bz, &params)
-	return params
-}
-
-// SetParams sets all module parameters.
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) error {
-	store := ctx.KVStore(k.state.key)
-	bz, err := k.cdc.Marshal(&params)
-	if err != nil {
-		return err
-	}
-	store.Set(types.ParamsKey, bz)
-	return nil
+	params, _ = k.Params.Get(ctx)
+	return
 }